stageflow_processor/requirements.txt
stageflow-core>=0.9.1
pydantic>=2.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
structlog>=24.0.0
